WI #2735 Refactor completion after TO
