+++ TypeCobol.LanguageServer.Test/LSRTests/CompletionWithNestedProgram/input/CompletionWithNestedProgram.tlsp
@@ -179,7 +179,7 @@
     },
     {
       "category": 1,
-      "message": "{\"jsonrpc\":\"2.0\",\"id\":\"7\",\"result\":{\"isIncomplete\":false,\"items\":[{\"label\":\"var2 (Numeric) (var2)\",\"kind\":6,\"insertText\":\"var2\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"YYYY (Numeric) (groupA-1::groupA-2::varA::YYYY)\",\"kind\":6,\"insertText\":\"groupA-1::groupA-2::varA::YYYY\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"YYYY (Numeric) (groupB-1::groupB-2::varB::YYYY)\",\"kind\":6,\"insertText\":\"groupB-1::groupB-2::varB::YYYY\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"MM (Numeric) (groupA-1::groupA-2::varA::MM)\",\"kind\":6,\"insertText\":\"groupA-1::groupA-2::varA::MM\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"MM (Numeric) (groupB-1::groupB-2::varB::MM)\",\"kind\":6,\"insertText\":\"groupB-1::groupB-2::varB::MM\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"DD (Numeric) (groupA-1::groupA-2::varA::DD)\",\"kind\":6,\"insertText\":\"groupA-1::groupA-2::varA::DD\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}},{\"label\":\"DD (Numeric) (groupB-1::groupB-2::varB::DD)\",\"kind\":6,\"insertText\":\"groupB-1::groupB-2::varB::DD\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}}]}}"
+      "message": "{\"jsonrpc\":\"2.0\",\"id\":\"7\",\"result\":{\"isIncomplete\":false,\"items\":[{\"label\":\"var2 (Numeric) (var2)\",\"kind\":6,\"insertText\":\"var2\",\"data\":{\"start\":{\"line\":38,\"character\":21},\"end\":{\"line\":38,\"character\":22}}}]}}"
     },
     {
       "category": 0,
